+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support index accessor for `json` function
+- Support `managementGroupResourceId` function
 
 ## [0.3.6] - 2022-05-21